fhirpipe_clean -> fhirpipe
