+++ fhirpipe/__init__.py
@@ -1,11 +1,13 @@
-from fhirpipe import parse  # noqa
-from fhirpipe import load  # noqa
-from fhirpipe import scripts  # noqa
-from fhirpipe import write  # noqa
+import yaml
+import logging
 
 global_config = None
 
 
-def set_global_config(new_config):
+def set_global_config(config_path):
     global global_config
-    global_config = new_config
+    with open(config_path, "r") as stream:
+        try:
+            global_config = yaml.safe_load(stream)
+        except yaml.YAMLError as exc:
+            logging.error(exc)
