--- fhirpipe/load/__init__.py
@@ -1,3 +1,0 @@
-from fhirpipe.load import sql  # noqa
-from fhirpipe.load import graphql  # noqa
-from fhirpipe.load import fhirstore  # noqa
