--- fhirpipe/parse/__init__.py
@@ -1,2 +1,0 @@
-from fhirpipe.parse import fhir  # noqa
-from fhirpipe.parse import sql  # noqa
