--- fhirpipe/write/__init__.py
@@ -1,1 +1,0 @@
-from fhirpipe.write.log import log  # noqa
