--- fhirpipe_clean/cli/__init__.py
@@ -1,14 +1,0 @@
-from .parse_args import parse_args  # noqa
-
-WELCOME_MSG = """
- (        )  (    (      (    (    (         
- )\ )  ( /(  )\ ) )\ )   )\ ) )\ ) )\ )      
-(()/(  )\())(()/((()/(  (()/((()/((()/( (    
- /(_))((_)\  /(_))/(_))  /(_))/(_))/(_)))\   
-(_))_| _((_)(_)) (_))   (_)) (_)) (_)) ((_)  
-| |_  | || ||_ _|| _ \  | _ \|_ _|| _ \| __| 
-| __| | __ | | | |   /  |  _/ | | |  _/| _|  
-|_|   |_||_||___||_|_\  |_|  |___||_|  |___| 
-
----------------- Arkhn Core ----------------
-"""  # noqa
