--- fhirpipe_clean/scripts/__init__.py
@@ -1,1 +1,0 @@
-from .standard import get_script  # noqa
