--- fhirpipe_clean/scripts/utils.py
@@ -1,2 +1,0 @@
-def is_empty(value):
-    return value is None or value == "NaN" or value == "" or value in "                        "
